requirements.txt
# appnope==0.1.3
asgiref==3.7.2
# asttokens==2.2.1
# backcall==0.2.0
bcrypt==4.0.1
Brotli==1.0.9
# cachetools==4.2.4
# certifi==2023.7.22
# charset-normalizer==3.2.0
# comm==0.1.3
# debugpy==1.6.7
# decorator==5.1.1
dj-database-url==2.0.0
Django==3.2.15
django-import-export==3.3.1
# django-googledrive-storage==1.6.0
# executing==1.2.0
# google-api-core==2.10.2
# google-api-python-client==2.95.0
# google-auth==2.22.0
# google-auth-httplib2==0.1.0
# google-auth-oauthlib==1.0.0
# googleapis-common-protos==1.60.0
gunicorn==20.1.0
# httplib2==0.22.0
# idna==3.4
# importlib-metadata==6.8.0
# ipykernel==6.24.0
# ipython==8.12.2
# jedi==0.18.2
# jupyter_client==8.3.0
# jupyter_core==5.3.1
# matplotlib-inline==0.1.6
# nest-asyncio==1.5.6
# oauthlib==3.2.2
# packaging==23.1
# parso==0.8.3
passlib==1.7.4
# pexpect==4.8.0
# pickleshare==0.7.5
Pillow==9.5.0
# platformdirs==3.9.1
# prompt-toolkit==3.0.39
# protobuf==4.23.4
# psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
# pyzmq==25.1.0
# requests==2.31.0
# requests-oauthlib==1.3.1
# rsa==4.9
# six==1.16.0
sqlparse==0.4.4
# stack-data==0.6.2
# tornado==6.3.2
# traitlets==5.9.0
typing_extensions==4.6.3
# uritemplate==4.1.1
# urllib3==1.26.16
# wcwidth==0.2.6
whitenoise==6.5.0
# zipp==3.16.2
